requirements.txt
Flask==0.10.1
Flask-Cors==1.10.3
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.10.1
argparse==1.2.1
itsdangerous==0.24
six==1.9.0
wsgiref==0.1.2
